+++ app/src/main/kotlin/hu/tajti/themealapp/model/Ingredient.kt
@@ -3,5 +3,5 @@ package hu.tajti.themealapp.model
 class Ingredient(
     var ingredientName: String? = null,
     var unit: String? = null,
-    var amount: Int? = null
+    var amount: Double? = null
 )
